Write parallel-transport fingerprint in native-readable integer bytes
P16 job58745737.8 reads authenticated Fe304 PT source and finite-link validation successfully after equivalent uint8-to-int32 metadata conversion; physical payload unchanged. Evidence Fe19/05_dft_orbital_dense/failed_velocity_memory.log. Subsequent orbital allocation failure is separate.

+++ src/file_io/parallel_transport.py
@@ -524,7 +524,7 @@ def initialize_parallel_transport_artifact(
         io.write_attr(
             "wfn_fingerprint_utf8",
             np.frombuffer(
-                str(wfn_fingerprint).encode("utf-8"), dtype=np.uint8))
+                str(wfn_fingerprint).encode("utf-8"), dtype=np.uint8).astype(np.int32))
         io.write_attr("polar_rcond", np.float64(rcond))
         # Numeric convention stamps are SlabIO-readable on every backend.
         # 1 means the sole supported convention documented by this schema.
